feat: improve contact management
- add toasts
- improve styles
- refactor
