+++ src/locales/cs.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+5HgiN\":[\"Mazání ubytování\"],\"+Vmqfs\":[\"Profilová fotka \",[\"name\"]],\"+j94GR\":[\"O vašem ubytování\"],\"/DXTzN\":[\"Prosím, <0>založte nový</0>.\"],\"/NgYQS\":[\"Zadejte svého Solid poskytovatele nebo WebID\"],\"/QPS+F\":[\"Otevřít profil\"],\"/Rg4q2\":[\"Kontakty <0/>\"],\"/qBoT3\":[\"Vytvořit veřejný index typů:\"],\"01t3Qz\":[\"Nepodařilo se odeslat zprávu: \"],\"17JZAf\":[\"Vytvořit schránku:\"],\"1Lxtdt\":[\"kód\"],\"1rUvkC\":[\"Preparing Solid Pod\"],\"2FwEm8\":[\"Ověřování...\"],\"2LDW9H\":[\"kontakty\"],\"2ZJVX8\":[\"logo \",[\"name\"]],\"2n699w\":[\"Community does not have a group.\"],\"2r6kDW\":[\"Hospex data setup successful\"],\"42aUkb\":[\"Zadejte svou e-mailovou adresu:\"],\"47l1wN\":[\"Vytvořit soukromý index typů:\"],\"49hNay\":[\"Můžete si také <0>provozovat vlastní Pod</0>.\"],\"4Wck3W\":[\"Request failed: \",[\"status\"],\" \",[\"statusText\"],\" \",[\"url\"]],\"5lWFkC\":[\"Přihlásit se\"],\"5mmKJA\":[\"Nový rozhovor s \",[\"people\"]],\"65VN7K\":[\"Zájmy\"],\"6Kt8NA\":[\"Načítání profilu\"],\"6YtxFj\":[\"Jméno\"],\"6dc9lc\":[\"Zájem odstraněn\"],\"6pe0RU\":[\"Platnost přihlášení vypršela!\"],\"7VpFf+\":[\"Your \",[\"communityLabel\"],\" data storage is set up.\"],\"7hktsm\":[\"Getting started\"],\"8BwQeU\":[\"Dokončit nastavení\"],\"8kXsjC\":[\"Ahoj! Rád(a) bych tě přidal(a) do kontaktů!\"],\"9K7sK1\":[\"E-mailové upozornění bylo odesláno\"],\"9OpVZg\":[\"Edit profile\"],\"A3rd0M\":[\"Nepodařilo se odeslat Linked Data upozornění.\"],\"A94/5H\":[\"Solid Pod is prepared\"],\"ARpyUK\":[\"Nepodařilo se vytvořit kanál\"],\"ApEC55\":[\"Lokalita hostování (klikněte na <0/> nebo přetáhněte a přibližte mapu)\"],\"AywWH9\":[\"Veřejný index typů je nastaven.\"],\"B8AaMI\":[\"Toto pole je povinné\"],\"BJjAHC\":[\"mé ubytování\"],\"BLFgtR\":[\"Připojit se ke komunitě\"],\"BQTO6T\":[\"Kvůli bezpečí sdílejte pouze přibližnou polohu\"],\"BajMQF\":[\"Odeslat pozvánku ke kontaktu\"],\"Bn8bAf\":[\"Pracujeme na tom\"],\"Bpp0Bz\":[\"Introduce yourself — it builds trust\"],\"C++TGh\":[\"Zavřít náhled hostingu\"],\"CDTRqk\":[\"Již jste členem \",[\"communityLabel\"]],\"CMfhBr\":[\"Hospex není nastaven (nemělo by se to stát)\"],\"CiQd9C\":[\"Přidat ubytování\"],\"Ct2zYB\":[\"Joining community\"],\"DvnzIZ\":[\"Tento rozhovor můžete jen číst.\"],\"ER/wp5\":[\"Aktualizuji profil\"],\"EnJuK0\":[\"Rozhovory\"],\"F132R+\":[\"Ověřte prosím svůj e-mail\"],\"FIq1Ba\":[\"Později\"],\"FQjzSF\":[\"Připravte Pod\"],\"GE4F4y\":[\"Začít nový rozhovor\"],\"GbJuzd\":[[\"messageCount\",\"plural\",{\"one\":[\"(\",\"#\",\" nová)\"],\"few\":[\"(\",\"#\",\" nové)\"],\"other\":[\"(\",\"#\",\" nových)\"]}]],\"H2Pd5s\":[\"Soubor předvoleb je nastaven.\"],\"HBI6nY\":[\"Potřebujete pomoc?\"],\"HFwLuM\":[\"Tento rozhovor má zastaralý formát.\"],\"HSNeM4\":[\"Community Solid Pod is not available at the moment.\"],\"HZ3V75\":[\"Probíhá odesílání Solid upozornění\"],\"HecDqj\":[\"upravit profil\"],\"Hwv3lE\":[\"Přihlásit se do \",[\"communityName\"]],\"I6Q2R9\":[\"Verification email sent\"],\"IAssyr\":[\"Please wait for everything to load.\"],\"IJbwnu\":[\"O mně\"],\"IYqOgL\":[\"Pozvěte prosím pouze lidi, které osobně znáte.\"],\"JlFcis\":[\"Odeslat\"],\"JmZ/+d\":[\"Finish\"],\"KBHVBP\":[\"Připojte se k \",[\"communityName\"]],\"KhBSE/\":[\"Invitation\"],\"KuVpz+\":[\"Unread\"],\"LShGrM\":[\"zdůrazněné logo \",[\"name\"]],\"M/R9wy\":[\"Vytváření zprávy\"],\"MHrjPM\":[\"Název\"],\"MPWsQ5\":[\"Veřejný index typů není nastaven\"],\"MYOoIS\":[\"Potvrzovací e-mail byl odeslán na adresu \",[\"sentToEmail\"],\".<0/>Zkontrolujte prosím svou schránku (i složku nevyžádané pošty) a postupujte podle pokynů.\"],\"N0mEtw\":[\"Vypadá to, že tady není žádný rozhovor.\"],\"NLJ18A\":[\"hospex kontejner nebyl nalezen (možná příliš brzy?)\"],\"NWQASA\":[\"E-mailové upozornění bylo odesláno.\"],\"NqP6cw\":[\"Zprávy s <0/>\"],\"O75yBm\":[\"načítání zájmů...\"],\"OjgFFJ\":[\"Prosím, přesuňte mapu na místo vašeho hostování\"],\"PS1p9R\":[\"Něco se pokazilo.\\nChyba: \",[\"e\"]],\"PUl3lo\":[\"Ubytování aktualizováno\"],\"PfxFCG\":[\"No hospex document for community\"],\"PmxYBp\":[\"Probíhá odesílání e-mailového upozornění\"],\"Pygpzd\":[\"Nepodařilo se nám přesměrovat na poskytovatele (\",[\"oidcIssuer\"],\").\\nZkontrolujte, zda jste zadali správné WebID nebo OIDCIssuer, nebo jestli není nedostupný.\\n\\nDůvod: \",[\"message\"],\"\\n\\nChyba: \",[\"errorString\"]],\"Q1iVjv\":[\"např. solidcommunity.net\"],\"QlfR57\":[\"Ubytování vytvořeno\"],\"QwanTV\":[\"Welcome to \",[\"communityLabel\"],\"!\"],\"Rad1+W\":[\"<0>Ozvěte se nám v našich projektových prostorech</0> nebo napište na <1>„support zavináč solidcouch tečka org\\\"</1>.\"],\"S48xcO\":[\"čeká\"],\"S8rY67\":[\"Odstraňování zájmu\"],\"SH3fsl\":[\"Preparing and sending verification email\"],\"TFYtVq\":[\"Přidat osobu do kontaktů\"],\"TQ1cfc\":[\"Accommodation not found.\"],\"V6us8s\":[\"Linked Data upozornění bylo odesláno.\"],\"VM9oSF\":[\"Vaše e-mailová upozornění jsou nastavena.\"],\"VXHB7t\":[\"Find people to stay with\"],\"WKVrQ3\":[\"Vyberte poskytovatele Podu:\"],\"WO3k4H\":[\"zpracování\"],\"Wixo6f\":[\"Opravdu chcete smazat ubytování?\"],\"XCH5Qf\":[\"Něco není vyřešeno (neočekávaná chyba)\"],\"XQAIBA\":[\"Neočekávaný stavový kód od indexační služby\"],\"XuPkTz\":[\"Skip getting started\"],\"XwS+Aq\":[\"Link existing community profile:\"],\"Y8WO6Z\":[\"Přidávání \",[\"label\"],\" do zájmů\"],\"YVkYKz\":[\"Welcome travellers to your home — if you can\"],\"Z0DAcX\":[\"Ubytování odebráno z indexační služby\"],\"Z3FXyt\":[\"Načítání...\"],\"Z3PAZP\":[\"Vaše přihlášení vypršelo. Přihlaste se, prosím, znovu.\"],\"ZDIydz\":[\"Get started\"],\"ZfmHL/\":[\"Choose your community data storage:\"],\"a8Qrhz\":[\"Something went wrong: \",[\"failure\"]],\"bC8dmZ\":[\"Create new profile for this community:\"],\"bGSJ3R\":[\"chybějící proměnné\"],\"bbvfbr\":[\"profil\"],\"bo2Xks\":[\"Načítání předchozích zpráv\"],\"bpK9mh\":[\"Confirm and Continue\"],\"cnGeoo\":[\"Smazat\"],\"d/3T+G\":[\"Schránka je nastavena.\"],\"dEgA5A\":[\"Zrušit\"],\"dIY3v5\":[\"Community joined\"],\"dJsK6G\":[\"Kontakt již existuje\"],\"dWGab4\":[\"Toto je začátek vaší konverzace\"],\"dg8/vM\":[\"Poslat zprávu…\"],\"e0IFR+\":[\"Přepnout na světlé téma\"],\"ePK91l\":[\"Upravit\"],\"eWlz27\":[\"Napsat zprávu\"],\"ei1Uj7\":[\"Soukromý index typu je nastaven.\"],\"f7fqUX\":[\"Ubytování aktualizováno v indexační službě\"],\"f8jrkd\":[\"více\"],\"fMnMgg\":[\"Soukromý index typu není nastaven (nemělo by se to stát).\"],\"fo1jSY\":[\"Schránka není nastavena\"],\"g3UF2V\":[\"Přijmout\"],\"gHFvtg\":[\"zprávy\"],\"gbbkE7\":[\"Ignorovat\"],\"hHo0hg\":[\"Nabídka ubytování od \",[\"offeredByIdentifier\"]],\"hXzOVo\":[\"Next\"],\"heRuNl\":[\"E-mailová upozornění nejsou pro tuto komunitu dostupná.\"],\"hyxB5M\":[\"Join request sent\"],\"iKT85g\":[\"Abyste se mohli připojit k \",[\"communityName\"],\", potřebujete <0>Solid Pod</0> — místo pro vaše data.\"],\"iRVV2L\":[\"Moje nabídky ubytování\"],\"ioBnUo\":[\"No storage found\"],\"jD5ug7\":[\"Nepodařilo se odeslat e-mailové upozornění.\"],\"jevYpv\":[\"Vytvořit soubor předvoleb:\"],\"jjIi3z\":[\"Ještě nemáte Solid Pod?\"],\"jlqRFZ\":[\"cestovat\"],\"jx1+4B\":[[\"label\"],\" byl přidán do zájmů\"],\"k4QqVn\":[\"Přidat \",[\"locale\"]],\"kH4vZi\":[\"Zpráva byla vytvořena\"],\"kH9S1u\":[\"Připojit se ke komunitě \",[\"communityLabel\"]],\"kf83Ld\":[\"Něco se nepovedlo.\"],\"lYCAyX\":[\"Nastavit e-mailová upozornění\"],\"maeAdU\":[\"načítání kontaktů\"],\"meH4FD\":[\"Přidat do mých kontaktů\"],\"mkeWa1\":[\"„\",[\"channel\"],\"“ není platný chatovací kanál\"],\"nwtY4N\":[\"Something went wrong\"],\"ocwjau\":[\"e-mailová adresa\"],\"pCyggK\":[\"Chatovací kanál byl vytvořen\"],\"piu7ZA\":[\"Zobrazit pozvánku ke kontaktu\"],\"qFkBv/\":[\"Solid upozornění bylo odesláno\"],\"qPXQY8\":[\"ověřování\"],\"qa9cy4\":[\"Fotka\"],\"r60ENi\":[\"Vytváření ubytování\"],\"sCXkJv\":[\"Nastavit oznámení\"],\"sRGBFr\":[\"schránka nebyla nalezena (možná příliš brzy?)\"],\"sa0+P5\":[\"Vyberte mou polohu\"],\"srPuxS\":[\"Profil aktualizován\"],\"t9VobG\":[\"Ubytování odstraněno\"],\"tEB6c1\":[\"hostit\"],\"tErrdj\":[\"Rozhovory s \",[\"people\"]],\"tfDRzk\":[\"Uložit\"],\"uLvjiZ\":[\"Přidat jazyk\"],\"udP0jg\":[\"Není dostupný žádný autentizovaný uživatel\"],\"uliB7J\":[\"Za \",[\"countdown\"],\"s znovu odeslat potvrzovací e-mail\"],\"uwiVz0\":[\"Pozvánka odeslána\"],\"v5eALp\":[\"Moje poloha\"],\"vJIeDW\":[\"Povězte ostatním o svém místě a hostování\"],\"vTkuq4\":[\"Probíhá oznámení indexační službě\"],\"vVEMd1\":[\"Hospex dokument nebyl nalezen\"],\"vp9yIB\":[\"Odeslat potvrzovací e-mail\"],\"vz3tpL\":[\"Ubytování přidáno do indexační služby\"],\"wRoTH5\":[\"Zprávy s \",[\"personLabel\"]],\"xAkrnZ\":[\"Schránka druhé osoby nebyla nalezena (pravděpodobně příliš brzy)\"],\"xGVfLh\":[\"Pokračovat\"],\"xdx7Sc\":[\"Aktualizuji ubytování\"],\"y3aU20\":[\"Uložit změny\"],\"yNR1Vx\":[\"Přepnout na tmavé téma\"],\"yQE2r9\":[\"Načítání\"],\"yXo/Y0\":[\"Postaveno na <0>SolidCouch</0> v\",[\"version\"],\" (<1>\",[\"commitHashShort\"],\"</1>)\"],\"ydgdD7\":[\"odhlásit se\"],\"yz7wBu\":[\"Zavřít\"],\"z8F4/3\":[\"Váš e-mail\"],\"zZiCw4\":[\"Setting up hospex data\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+5HgiN\":[\"Mazání ubytování\"],\"+Vmqfs\":[\"Profilová fotka \",[\"name\"]],\"+j94GR\":[\"O vašem ubytování\"],\"/DXTzN\":[\"Prosím, <0>založte nový</0>.\"],\"/NgYQS\":[\"Zadejte svého Solid poskytovatele nebo WebID\"],\"/QPS+F\":[\"Otevřít profil\"],\"/Rg4q2\":[\"Kontakty <0/>\"],\"/ZSU/m\":[\"Contact confirmed\"],\"/qBoT3\":[\"Vytvořit veřejný index typů:\"],\"01t3Qz\":[\"Nepodařilo se odeslat zprávu: \"],\"17JZAf\":[\"Vytvořit schránku:\"],\"1Lxtdt\":[\"kód\"],\"1i+H23\":[\"Ignoring contact request\"],\"1rUvkC\":[\"Preparing Solid Pod\"],\"2FwEm8\":[\"Ověřování...\"],\"2LDW9H\":[\"kontakty\"],\"2ZJVX8\":[\"logo \",[\"name\"]],\"2n699w\":[\"Community does not have a group.\"],\"2r6kDW\":[\"Hospex data setup successful\"],\"3pqzUK\":[\"Contact added.\"],\"42aUkb\":[\"Zadejte svou e-mailovou adresu:\"],\"47l1wN\":[\"Vytvořit soukromý index typů:\"],\"49hNay\":[\"Můžete si také <0>provozovat vlastní Pod</0>.\"],\"4Wck3W\":[\"Request failed: \",[\"status\"],\" \",[\"statusText\"],\" \",[\"url\"]],\"5lWFkC\":[\"Přihlásit se\"],\"5mmKJA\":[\"Nový rozhovor s \",[\"people\"]],\"65VN7K\":[\"Zájmy\"],\"6Kt8NA\":[\"Načítání profilu\"],\"6YtxFj\":[\"Jméno\"],\"6dc9lc\":[\"Zájem odstraněn\"],\"6pe0RU\":[\"Platnost přihlášení vypršela!\"],\"7VpFf+\":[\"Your \",[\"communityLabel\"],\" data storage is set up.\"],\"7hktsm\":[\"Getting started\"],\"8BwQeU\":[\"Dokončit nastavení\"],\"8kXsjC\":[\"Ahoj! Rád(a) bych tě přidal(a) do kontaktů!\"],\"9K7sK1\":[\"E-mailové upozornění bylo odesláno\"],\"9OpVZg\":[\"Edit profile\"],\"A3rd0M\":[\"Nepodařilo se odeslat Linked Data upozornění.\"],\"A94/5H\":[\"Solid Pod is prepared\"],\"ARpyUK\":[\"Nepodařilo se vytvořit kanál\"],\"ApEC55\":[\"Lokalita hostování (klikněte na <0/> nebo přetáhněte a přibližte mapu)\"],\"AywWH9\":[\"Veřejný index typů je nastaven.\"],\"B8AaMI\":[\"Toto pole je povinné\"],\"BJjAHC\":[\"mé ubytování\"],\"BLFgtR\":[\"Připojit se ke komunitě\"],\"BQTO6T\":[\"Kvůli bezpečí sdílejte pouze přibližnou polohu\"],\"BajMQF\":[\"Odeslat pozvánku ke kontaktu\"],\"Bn8bAf\":[\"Pracujeme na tom\"],\"Bpp0Bz\":[\"Introduce yourself — it builds trust\"],\"C++TGh\":[\"Zavřít náhled hostingu\"],\"CDTRqk\":[\"Již jste členem \",[\"communityLabel\"]],\"CMfhBr\":[\"Hospex není nastaven (nemělo by se to stát)\"],\"CiQd9C\":[\"Přidat ubytování\"],\"Ct2zYB\":[\"Joining community\"],\"DvnzIZ\":[\"Tento rozhovor můžete jen číst.\"],\"ER/wp5\":[\"Aktualizuji profil\"],\"EbjjWU\":[\"Confirming contact\"],\"EnJuK0\":[\"Rozhovory\"],\"F132R+\":[\"Ověřte prosím svůj e-mail\"],\"FIq1Ba\":[\"Později\"],\"FQjzSF\":[\"Připravte Pod\"],\"GE4F4y\":[\"Začít nový rozhovor\"],\"GbJuzd\":[[\"messageCount\",\"plural\",{\"one\":[\"(\",\"#\",\" nová)\"],\"few\":[\"(\",\"#\",\" nové)\"],\"other\":[\"(\",\"#\",\" nových)\"]}]],\"H2Pd5s\":[\"Soubor předvoleb je nastaven.\"],\"HBI6nY\":[\"Potřebujete pomoc?\"],\"HFwLuM\":[\"Tento rozhovor má zastaralý formát.\"],\"HSNeM4\":[\"Community Solid Pod is not available at the moment.\"],\"HZ3V75\":[\"Probíhá odesílání Solid upozornění\"],\"HecDqj\":[\"upravit profil\"],\"Hwv3lE\":[\"Přihlásit se do \",[\"communityName\"]],\"I6Q2R9\":[\"Verification email sent\"],\"IAssyr\":[\"Please wait for everything to load.\"],\"IJbwnu\":[\"O mně\"],\"IYqOgL\":[\"Pozvěte prosím pouze lidi, které osobně znáte.\"],\"JlFcis\":[\"Odeslat\"],\"JmZ/+d\":[\"Finish\"],\"KBHVBP\":[\"Připojte se k \",[\"communityName\"]],\"KhBSE/\":[\"Invitation\"],\"KuVpz+\":[\"Unread\"],\"LShGrM\":[\"zdůrazněné logo \",[\"name\"]],\"LpmI27\":[\"Adding contact.\"],\"M/R9wy\":[\"Vytváření zprávy\"],\"MHrjPM\":[\"Název\"],\"MPWsQ5\":[\"Veřejný index typů není nastaven\"],\"MYOoIS\":[\"Potvrzovací e-mail byl odeslán na adresu \",[\"sentToEmail\"],\".<0/>Zkontrolujte prosím svou schránku (i složku nevyžádané pošty) a postupujte podle pokynů.\"],\"N0mEtw\":[\"Vypadá to, že tady není žádný rozhovor.\"],\"NLJ18A\":[\"hospex kontejner nebyl nalezen (možná příliš brzy?)\"],\"NWQASA\":[\"E-mailové upozornění bylo odesláno.\"],\"NqP6cw\":[\"Zprávy s <0/>\"],\"O75yBm\":[\"načítání zájmů...\"],\"OjgFFJ\":[\"Prosím, přesuňte mapu na místo vašeho hostování\"],\"PS1p9R\":[\"Něco se pokazilo.\\nChyba: \",[\"e\"]],\"PUl3lo\":[\"Ubytování aktualizováno\"],\"PfxFCG\":[\"No hospex document for community\"],\"PmxYBp\":[\"Probíhá odesílání e-mailového upozornění\"],\"Pygpzd\":[\"Nepodařilo se nám přesměrovat na poskytovatele (\",[\"oidcIssuer\"],\").\\nZkontrolujte, zda jste zadali správné WebID nebo OIDCIssuer, nebo jestli není nedostupný.\\n\\nDůvod: \",[\"message\"],\"\\n\\nChyba: \",[\"errorString\"]],\"Q1iVjv\":[\"např. solidcommunity.net\"],\"QlfR57\":[\"Ubytování vytvořeno\"],\"QwanTV\":[\"Welcome to \",[\"communityLabel\"],\"!\"],\"Rad1+W\":[\"<0>Ozvěte se nám v našich projektových prostorech</0> nebo napište na <1>„support zavináč solidcouch tečka org\\\"</1>.\"],\"S48xcO\":[\"čeká\"],\"S8rY67\":[\"Odstraňování zájmu\"],\"SH3fsl\":[\"Preparing and sending verification email\"],\"TFYtVq\":[\"Přidat osobu do kontaktů\"],\"TQ1cfc\":[\"Accommodation not found.\"],\"V6us8s\":[\"Linked Data upozornění bylo odesláno.\"],\"VM9oSF\":[\"Vaše e-mailová upozornění jsou nastavena.\"],\"VXHB7t\":[\"Find people to stay with\"],\"VhYqBG\":[\"Contact request from <0/>\"],\"WKVrQ3\":[\"Vyberte poskytovatele Podu:\"],\"WO3k4H\":[\"zpracování\"],\"Wixo6f\":[\"Opravdu chcete smazat ubytování?\"],\"XCH5Qf\":[\"Něco není vyřešeno (neočekávaná chyba)\"],\"XQAIBA\":[\"Neočekávaný stavový kód od indexační služby\"],\"XuPkTz\":[\"Skip getting started\"],\"XwS+Aq\":[\"Link existing community profile:\"],\"Y8WO6Z\":[\"Přidávání \",[\"label\"],\" do zájmů\"],\"YVkYKz\":[\"Welcome travellers to your home — if you can\"],\"Z0DAcX\":[\"Ubytování odebráno z indexační služby\"],\"Z3FXyt\":[\"Načítání...\"],\"Z3PAZP\":[\"Vaše přihlášení vypršelo. Přihlaste se, prosím, znovu.\"],\"ZDIydz\":[\"Get started\"],\"ZfmHL/\":[\"Choose your community data storage:\"],\"a8Qrhz\":[\"Something went wrong: \",[\"failure\"]],\"bC8dmZ\":[\"Create new profile for this community:\"],\"bGSJ3R\":[\"chybějící proměnné\"],\"bbvfbr\":[\"profil\"],\"bo2Xks\":[\"Načítání předchozích zpráv\"],\"bpK9mh\":[\"Confirm and Continue\"],\"cnGeoo\":[\"Smazat\"],\"d/3T+G\":[\"Schránka je nastavena.\"],\"dEgA5A\":[\"Zrušit\"],\"dIY3v5\":[\"Community joined\"],\"dJsK6G\":[\"Kontakt již existuje\"],\"dWGab4\":[\"Toto je začátek vaší konverzace\"],\"dg8/vM\":[\"Poslat zprávu…\"],\"e0IFR+\":[\"Přepnout na světlé téma\"],\"ePK91l\":[\"Upravit\"],\"eWlz27\":[\"Napsat zprávu\"],\"ei1Uj7\":[\"Soukromý index typu je nastaven.\"],\"f7fqUX\":[\"Ubytování aktualizováno v indexační službě\"],\"f8jrkd\":[\"více\"],\"fMnMgg\":[\"Soukromý index typu není nastaven (nemělo by se to stát).\"],\"fo1jSY\":[\"Schránka není nastavena\"],\"g3UF2V\":[\"Přijmout\"],\"gHFvtg\":[\"zprávy\"],\"gbbkE7\":[\"Ignorovat\"],\"hHo0hg\":[\"Nabídka ubytování od \",[\"offeredByIdentifier\"]],\"hXzOVo\":[\"Next\"],\"heRuNl\":[\"E-mailová upozornění nejsou pro tuto komunitu dostupná.\"],\"hyxB5M\":[\"Join request sent\"],\"iKT85g\":[\"Abyste se mohli připojit k \",[\"communityName\"],\", potřebujete <0>Solid Pod</0> — místo pro vaše data.\"],\"iRVV2L\":[\"Moje nabídky ubytování\"],\"ioBnUo\":[\"No storage found\"],\"jD5ug7\":[\"Nepodařilo se odeslat e-mailové upozornění.\"],\"jRqiIC\":[\"Contact request ignored\"],\"jevYpv\":[\"Vytvořit soubor předvoleb:\"],\"jjIi3z\":[\"Ještě nemáte Solid Pod?\"],\"jlqRFZ\":[\"cestovat\"],\"jx1+4B\":[[\"label\"],\" byl přidán do zájmů\"],\"k4QqVn\":[\"Přidat \",[\"locale\"]],\"kH4vZi\":[\"Zpráva byla vytvořena\"],\"kH9S1u\":[\"Připojit se ke komunitě \",[\"communityLabel\"]],\"kf83Ld\":[\"Něco se nepovedlo.\"],\"lYCAyX\":[\"Nastavit e-mailová upozornění\"],\"maeAdU\":[\"načítání kontaktů\"],\"meH4FD\":[\"Přidat do mých kontaktů\"],\"mkeWa1\":[\"„\",[\"channel\"],\"“ není platný chatovací kanál\"],\"nwtY4N\":[\"Something went wrong\"],\"ocwjau\":[\"e-mailová adresa\"],\"pCyggK\":[\"Chatovací kanál byl vytvořen\"],\"pY95XS\":[\"<0/> wrote you:\"],\"piu7ZA\":[\"Zobrazit pozvánku ke kontaktu\"],\"qFkBv/\":[\"Solid upozornění bylo odesláno\"],\"qPXQY8\":[\"ověřování\"],\"qa9cy4\":[\"Fotka\"],\"r60ENi\":[\"Vytváření ubytování\"],\"sCXkJv\":[\"Nastavit oznámení\"],\"sRGBFr\":[\"schránka nebyla nalezena (možná příliš brzy?)\"],\"sa0+P5\":[\"Vyberte mou polohu\"],\"srPuxS\":[\"Profil aktualizován\"],\"t9VobG\":[\"Ubytování odstraněno\"],\"tEB6c1\":[\"hostit\"],\"tErrdj\":[\"Rozhovory s \",[\"people\"]],\"tfDRzk\":[\"Uložit\"],\"uLvjiZ\":[\"Přidat jazyk\"],\"udP0jg\":[\"Není dostupný žádný autentizovaný uživatel\"],\"uliB7J\":[\"Za \",[\"countdown\"],\"s znovu odeslat potvrzovací e-mail\"],\"uwiVz0\":[\"Pozvánka odeslána\"],\"v5eALp\":[\"Moje poloha\"],\"vJIeDW\":[\"Povězte ostatním o svém místě a hostování\"],\"vTkuq4\":[\"Probíhá oznámení indexační službě\"],\"vVEMd1\":[\"Hospex dokument nebyl nalezen\"],\"vp9yIB\":[\"Odeslat potvrzovací e-mail\"],\"vz3tpL\":[\"Ubytování přidáno do indexační služby\"],\"wRoTH5\":[\"Zprávy s \",[\"personLabel\"]],\"xAkrnZ\":[\"Schránka druhé osoby nebyla nalezena (pravděpodobně příliš brzy)\"],\"xGVfLh\":[\"Pokračovat\"],\"xdx7Sc\":[\"Aktualizuji ubytování\"],\"y3aU20\":[\"Uložit změny\"],\"yNR1Vx\":[\"Přepnout na tmavé téma\"],\"yQE2r9\":[\"Načítání\"],\"yXo/Y0\":[\"Postaveno na <0>SolidCouch</0> v\",[\"version\"],\" (<1>\",[\"commitHashShort\"],\"</1>)\"],\"ydgdD7\":[\"odhlásit se\"],\"yz7wBu\":[\"Zavřít\"],\"z8F4/3\":[\"Váš e-mail\"],\"zZiCw4\":[\"Setting up hospex data\"]}")as Messages;
