+++ src/locales/en.po
@@ -26,6 +26,10 @@ msgstr "{label} added to interests"
 msgid "{messageCount, plural, one {(# new)} other {(# new)}}"
 msgstr "{messageCount, plural, one {(# new)} other {(# new)}}"
 
+#: src/pages/Profile/ProcessContactRequest.tsx:90
+msgid "<0/> wrote you:"
+msgstr "<0/> wrote you:"
+
 #: src/components/Join/Join.tsx:115
 msgid "<0>Reach us in our project spaces</0> or write us email to <1>\"support at solidcouch dot org\"</1>."
 msgstr "<0>Reach us in our project spaces</0> or write us email to <1>\"support at solidcouch dot org\"</1>."
@@ -42,7 +46,7 @@ msgstr "About me"
 msgid "About your hosting"
 msgstr "About your hosting"
 
-#: src/pages/Profile/ManageContact.tsx:99
+#: src/pages/Profile/ProcessContactRequest.tsx:106
 msgid "Accept"
 msgstr "Accept"
 
@@ -90,17 +94,21 @@ msgstr "Add a language"
 msgid "Add Accommodation"
 msgstr "Add Accommodation"
 
-#: src/pages/Profile/ManageContact.tsx:181
+#: src/pages/Profile/AddContactForm.tsx:25
 msgid "Add person as a contact"
 msgstr "Add person as a contact"
 
-#: src/pages/Profile/ManageContact.tsx:153
+#: src/pages/Profile/AddContact.tsx:73
 msgid "Add to my contacts"
 msgstr "Add to my contacts"
 
 #: src/pages/EditProfile/EditInterests.tsx:64
 msgid "Adding {label} to interests"
 msgstr "Adding {label} to interests"
+
+#: src/pages/Profile/AddContact.tsx:50
+msgid "Adding contact."
+msgstr "Adding contact."
 
 #: src/pages/Contacts.tsx:20
 #: src/pages/ProfileRedirect.tsx:13
@@ -113,7 +121,7 @@ msgstr "Authenticating..."
 
 #: src/components/Accommodation/AccommodationForm.tsx:159
 #: src/components/LocaleTextInput/LocaleTextInput.tsx:188
-#: src/pages/Profile/ManageContact.tsx:192
+#: src/pages/Profile/AddContactForm.tsx:38
 msgid "Cancel"
 msgstr "Cancel"
 
@@ -164,11 +172,31 @@ msgstr "Community Solid Pod is not available at the moment."
 msgid "Confirm and Continue"
 msgstr "Confirm and Continue"
 
-#: src/pages/Profile/ManageContact.tsx:35
+#: src/pages/Profile/ProcessContactRequest.tsx:55
+msgid "Confirming contact"
+msgstr "Confirming contact"
+
+#: src/pages/Profile/AddContact.tsx:51
+msgid "Contact added."
+msgstr "Contact added."
+
+#: src/pages/Profile/ProcessContactRequest.tsx:56
+msgid "Contact confirmed"
+msgstr "Contact confirmed"
+
+#: src/pages/Profile/ManageContact.tsx:25
 msgid "Contact exists"
 msgstr "Contact exists"
 
-#: src/pages/Profile/ManageContact.tsx:38
+#: src/pages/Profile/ProcessContactRequest.tsx:83
+msgid "Contact request from <0/>"
+msgstr "Contact request from <0/>"
+
+#: src/pages/Profile/ProcessContactRequest.tsx:68
+msgid "Contact request ignored"
+msgstr "Contact request ignored"
+
+#: src/pages/Profile/ManageContact.tsx:28
 msgid "Contact request sent"
 msgstr "Contact request sent"
 
@@ -319,12 +347,12 @@ msgstr "Get started"
 msgid "Getting started"
 msgstr "Getting started"
 
-#: src/pages/Profile/ManageContact.tsx:171
+#: src/pages/Profile/AddContactForm.tsx:15
 msgid "Hi! I'd like to add you as a contact!"
 msgstr "Hi! I'd like to add you as a contact!"
 
-#: src/pages/Profile/ManageContact.tsx:70
-#: src/pages/Profile/ManageContact.tsx:127
+#: src/pages/Profile/AddContact.tsx:35
+#: src/pages/Profile/ProcessContactRequest.tsx:46
 msgid "hospex container not found (too soon?)"
 msgstr "hospex container not found (too soon?)"
 
@@ -349,9 +377,13 @@ msgid "Hosting location (click <0/> or drag & zoom map)"
 msgstr "Hosting location (click <0/> or drag & zoom map)"
 
 #: src/pages/messages/Messages.tsx:475
-#: src/pages/Profile/ManageContact.tsx:102
+#: src/pages/Profile/ProcessContactRequest.tsx:99
 msgid "Ignore"
 msgstr "Ignore"
+
+#: src/pages/Profile/ProcessContactRequest.tsx:67
+msgid "Ignoring contact request"
+msgstr "Ignoring contact request"
 
 #: src/pages/HospexSetup/NotificationsSetup.tsx:216
 msgid "Inbox is not set up"
@@ -361,7 +393,7 @@ msgstr "Inbox is not set up"
 msgid "Inbox is set up."
 msgstr "Inbox is set up."
 
-#: src/pages/Profile/ManageContact.tsx:129
+#: src/pages/Profile/AddContact.tsx:39
 msgid "inbox not found (too soon?)"
 msgstr "inbox not found (too soon?)"
 
@@ -435,7 +467,7 @@ msgstr "Loading previous messages"
 #: src/App.tsx:54
 #: src/pages/messages-old/MessagesOld.tsx:213
 #: src/pages/MyOffers.tsx:25
-#: src/pages/Profile/ManageContact.tsx:26
+#: src/pages/Profile/ManageContact.tsx:16
 #: src/pages/SearchHosts/AccommodationInfo.tsx:54
 msgid "Loading..."
 msgstr "Loading..."
@@ -533,7 +565,7 @@ msgstr "Photo"
 msgid "Please <0>start a new one</0>."
 msgstr "Please <0>start a new one</0>."
 
-#: src/pages/Profile/ManageContact.tsx:184
+#: src/pages/Profile/AddContactForm.tsx:28
 msgid "Please invite only people you know personally."
 msgstr "Please invite only people you know personally."
 
@@ -622,7 +654,7 @@ msgstr "Save"
 msgid "Save changes"
 msgstr "Save changes"
 
-#: src/pages/Profile/ManageContact.tsx:90
+#: src/pages/Profile/ProcessContactRequest.tsx:77
 msgid "See contact invitation"
 msgstr "See contact invitation"
 
@@ -644,7 +676,7 @@ msgstr "Send a message…"
 msgid "Send Confirmation Email"
 msgstr "Send Confirmation Email"
 
-#: src/pages/Profile/ManageContact.tsx:189
+#: src/pages/Profile/AddContactForm.tsx:41
 msgid "Send contact invitation"
 msgstr "Send contact invitation"
