+++ src/locales/en.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+5HgiN\":[\"Deleting accommodation\"],\"+Vmqfs\":[\"Profile photo of \",[\"name\"]],\"+j94GR\":[\"About your hosting\"],\"/DXTzN\":[\"Please <0>start a new one</0>.\"],\"/NgYQS\":[\"Choose your Solid provider or WebID\"],\"/QPS+F\":[\"Open profile\"],\"/Rg4q2\":[\"Contacts of <0/>\"],\"/qBoT3\":[\"Create public type index:\"],\"01t3Qz\":[\"Sending message failed: \"],\"17JZAf\":[\"Create inbox:\"],\"1Lxtdt\":[\"code\"],\"1rUvkC\":[\"Preparing Solid Pod\"],\"2FwEm8\":[\"Authenticating...\"],\"2LDW9H\":[\"contacts\"],\"2ZJVX8\":[\"logo of \",[\"name\"]],\"2n699w\":[\"Community does not have a group.\"],\"2r6kDW\":[\"Hospex data setup successful\"],\"42aUkb\":[\"Enter your email address:\"],\"47l1wN\":[\"Create private type index:\"],\"49hNay\":[\"You can also <0>host your own Pod</0>.\"],\"4Wck3W\":[\"Request failed: \",[\"status\"],\" \",[\"statusText\"],\" \",[\"url\"]],\"5lWFkC\":[\"Sign in\"],\"5mmKJA\":[\"New conversation with \",[\"people\"]],\"65VN7K\":[\"Interests\"],\"6Kt8NA\":[\"Fetching profile\"],\"6YtxFj\":[\"Name\"],\"6dc9lc\":[\"Interest removed\"],\"6pe0RU\":[\"Session Expired!\"],\"7VpFf+\":[\"Your \",[\"communityLabel\"],\" data storage is set up.\"],\"7hktsm\":[\"Getting started\"],\"8BwQeU\":[\"Finish Setup\"],\"8kXsjC\":[\"Hi! I'd like to add you as a contact!\"],\"9K7sK1\":[\"Email notification was sent\"],\"9OpVZg\":[\"Edit profile\"],\"A3rd0M\":[\"Sending Linked Data Notification failed.\"],\"A94/5H\":[\"Solid Pod is prepared\"],\"ARpyUK\":[\"Creating channel failed\"],\"ApEC55\":[\"Hosting location (click <0/> or drag & zoom map)\"],\"AywWH9\":[\"Public type index is set up.\"],\"B8AaMI\":[\"This field is required\"],\"BJjAHC\":[\"my hosting\"],\"BLFgtR\":[\"Join Community\"],\"BQTO6T\":[\"For safety, share only approximate location\"],\"BajMQF\":[\"Send contact invitation\"],\"Bn8bAf\":[\"WIP\"],\"Bpp0Bz\":[\"Introduce yourself — it builds trust\"],\"C++TGh\":[\"Close hosting preview\"],\"CDTRqk\":[\"You are already a member of \",[\"communityLabel\"]],\"CMfhBr\":[\"Hospex not set up (should not happen)\"],\"CiQd9C\":[\"Add Accommodation\"],\"Ct2zYB\":[\"Joining community\"],\"DvnzIZ\":[\"You can only read this chat.\"],\"ER/wp5\":[\"Updating profile\"],\"EnJuK0\":[\"Conversations\"],\"F132R+\":[\"Please verify your email\"],\"FIq1Ba\":[\"Later\"],\"FQjzSF\":[\"Prepare Pod\"],\"GE4F4y\":[\"Start a new conversation\"],\"GbJuzd\":[[\"messageCount\",\"plural\",{\"one\":[\"(\",\"#\",\" new)\"],\"other\":[\"(\",\"#\",\" new)\"]}]],\"H2Pd5s\":[\"Preferences file is set up.\"],\"HBI6nY\":[\"Need help?\"],\"HFwLuM\":[\"This chat has outdated format.\"],\"HSNeM4\":[\"Community Solid Pod is not available at the moment.\"],\"HZ3V75\":[\"Sending Solid notification\"],\"HecDqj\":[\"edit profile\"],\"Hwv3lE\":[\"Sign in to \",[\"communityName\"]],\"I6Q2R9\":[\"Verification email sent\"],\"IAssyr\":[\"Please wait for everything to load.\"],\"IJbwnu\":[\"About me\"],\"IYqOgL\":[\"Please invite only people you know personally.\"],\"JlFcis\":[\"Send\"],\"JmZ/+d\":[\"Finish\"],\"KBHVBP\":[\"Join \",[\"communityName\"]],\"KhBSE/\":[\"Invitation\"],\"KuVpz+\":[\"Unread\"],\"LShGrM\":[\"focused logo of \",[\"name\"]],\"M/R9wy\":[\"Creating message\"],\"MHrjPM\":[\"Title\"],\"MPWsQ5\":[\"Public type index is not set up\"],\"MYOoIS\":[\"A confirmation email has been sent to \",[\"sentToEmail\"],\".<0/>Please check your inbox (and spam folder) and follow the instructions.\"],\"N0mEtw\":[\"There doesn't seem to be a chat here.\"],\"NLJ18A\":[\"hospex container not found (too soon?)\"],\"NWQASA\":[\"Email Notification was sent.\"],\"NqP6cw\":[\"Messages with <0/>\"],\"O75yBm\":[\"loading interests...\"],\"OjgFFJ\":[\"Please move map to your hosting location\"],\"PS1p9R\":[\"Something went wrong.\\nError: \",[\"e\"]],\"PUl3lo\":[\"Accommodation updated\"],\"PfxFCG\":[\"No hospex document for community\"],\"PmxYBp\":[\"Sending email notification\"],\"Pygpzd\":[\"We didn't succeed with redirecting to the issuer at \",[\"oidcIssuer\"],\".\\nHave you provided correct webId or OIDCIssuer? Or is it down?\\n\\nReason: \",[\"message\"],\"\\n\\nError: \",[\"errorString\"]],\"Q1iVjv\":[\"e.g. solidcommunity.net\"],\"QlfR57\":[\"Accommodation created\"],\"QwanTV\":[\"Welcome to \",[\"communityLabel\"],\"!\"],\"Rad1+W\":[\"<0>Reach us in our project spaces</0> or write us email to <1>\\\"support at solidcouch dot org\\\"</1>.\"],\"S48xcO\":[\"pending\"],\"S8rY67\":[\"Removing interest\"],\"SH3fsl\":[\"Preparing and sending verification email\"],\"TFYtVq\":[\"Add person as a contact\"],\"TQ1cfc\":[\"Accommodation not found.\"],\"V6us8s\":[\"Linked Data Notification was sent.\"],\"VM9oSF\":[\"Your email notifications are set up.\"],\"VXHB7t\":[\"Find people to stay with\"],\"WKVrQ3\":[\"Choose a Pod provider:\"],\"WO3k4H\":[\"process\"],\"Wixo6f\":[\"Do you really want to delete the accommodation?\"],\"XCH5Qf\":[\"Something is not resolved (unexpected error)\"],\"XQAIBA\":[\"Unexpected status code from indexing service\"],\"XuPkTz\":[\"Skip getting started\"],\"XwS+Aq\":[\"Link existing community profile:\"],\"Y8WO6Z\":[\"Adding \",[\"label\"],\" to interests\"],\"YVkYKz\":[\"Welcome travellers to your home — if you can\"],\"Z0DAcX\":[\"Accommodation removed from indexing service\"],\"Z3FXyt\":[\"Loading...\"],\"Z3PAZP\":[\"Your session has expired. Please sign in again.\"],\"ZDIydz\":[\"Get started\"],\"ZfmHL/\":[\"Choose your community data storage:\"],\"a8Qrhz\":[\"Something went wrong: \",[\"failure\"]],\"bC8dmZ\":[\"Create new profile for this community:\"],\"bGSJ3R\":[\"missing variables\"],\"bbvfbr\":[\"profile\"],\"bo2Xks\":[\"Loading previous messages\"],\"bpK9mh\":[\"Confirm and Continue\"],\"cnGeoo\":[\"Delete\"],\"d/3T+G\":[\"Inbox is set up.\"],\"dEgA5A\":[\"Cancel\"],\"dIY3v5\":[\"Community joined\"],\"dJsK6G\":[\"Contact exists\"],\"dWGab4\":[\"This is a start of your conversation\"],\"dg8/vM\":[\"Send a message…\"],\"e0IFR+\":[\"Switch to light theme\"],\"ePK91l\":[\"Edit\"],\"eWlz27\":[\"Write a message\"],\"ei1Uj7\":[\"Private type index is set up.\"],\"f7fqUX\":[\"Accommodation updated in indexing service\"],\"f8jrkd\":[\"more\"],\"fMnMgg\":[\"Private type index not set up (should not happen)\"],\"fo1jSY\":[\"Inbox is not set up\"],\"g3UF2V\":[\"Accept\"],\"gHFvtg\":[\"messages\"],\"gbbkE7\":[\"Ignore\"],\"hHo0hg\":[\"Accommodation offer from \",[\"offeredByIdentifier\"]],\"hXzOVo\":[\"Next\"],\"heRuNl\":[\"Email notifications are not available for this community.\"],\"hyxB5M\":[\"Join request sent\"],\"iKT85g\":[\"To join \",[\"communityName\"],\", you need a <0>Solid Pod</0> — a place to store your data.\"],\"iRVV2L\":[\"My Accommodation Offers\"],\"ioBnUo\":[\"No storage found\"],\"jD5ug7\":[\"Sending Email Notification failed.\"],\"jevYpv\":[\"Create preferences file:\"],\"jjIi3z\":[\"Don't have a Solid Pod yet?\"],\"jlqRFZ\":[\"travel\"],\"jx1+4B\":[[\"label\"],\" added to interests\"],\"k4QqVn\":[\"Add \",[\"locale\"]],\"kH4vZi\":[\"Message was created\"],\"kH9S1u\":[\"Join community \",[\"communityLabel\"]],\"kf83Ld\":[\"Something went wrong.\"],\"lYCAyX\":[\"Setup email notifications\"],\"maeAdU\":[\"fetching contacts\"],\"meH4FD\":[\"Add to my contacts\"],\"mkeWa1\":[\"\\\"\",[\"channel\"],\"\\\" is not a valid chat channel\"],\"nwtY4N\":[\"Something went wrong\"],\"ocwjau\":[\"email address\"],\"pCyggK\":[\"Chat channel was created\"],\"piu7ZA\":[\"See contact invitation\"],\"qFkBv/\":[\"Solid notification was sent\"],\"qPXQY8\":[\"authenticating\"],\"qa9cy4\":[\"Photo\"],\"r60ENi\":[\"Creating accommodation\"],\"sCXkJv\":[\"Set Up Notifications\"],\"sRGBFr\":[\"inbox not found (too soon?)\"],\"sa0+P5\":[\"Select my location\"],\"srPuxS\":[\"Profile updated\"],\"t9VobG\":[\"Accommodation deleted\"],\"tEB6c1\":[\"host\"],\"tErrdj\":[\"Conversations with \",[\"people\"]],\"tfDRzk\":[\"Save\"],\"uLvjiZ\":[\"Add a language\"],\"udP0jg\":[\"No authenticated user available\"],\"uliB7J\":[\"Resend Confirmation Email in \",[\"countdown\"],\"s\"],\"uwiVz0\":[\"Contact request sent\"],\"v5eALp\":[\"My location\"],\"vJIeDW\":[\"Tell others about your place and hosting\"],\"vTkuq4\":[\"Notifying indexing service\"],\"vVEMd1\":[\"Hospex document not found\"],\"vp9yIB\":[\"Send Confirmation Email\"],\"vz3tpL\":[\"Accommodation added to indexing service\"],\"wRoTH5\":[\"Messages with \",[\"personLabel\"]],\"xAkrnZ\":[\"Inbox of the other person not found (probably too soon)\"],\"xGVfLh\":[\"Continue\"],\"xdx7Sc\":[\"Updating accommodation\"],\"y3aU20\":[\"Save changes\"],\"yNR1Vx\":[\"Switch to dark theme\"],\"yQE2r9\":[\"Loading\"],\"yXo/Y0\":[\"Powered by <0>SolidCouch</0> v\",[\"version\"],\" (<1>\",[\"commitHashShort\"],\"</1>)\"],\"ydgdD7\":[\"sign out\"],\"yz7wBu\":[\"Close\"],\"z8F4/3\":[\"Your email\"],\"zZiCw4\":[\"Setting up hospex data\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+5HgiN\":[\"Deleting accommodation\"],\"+Vmqfs\":[\"Profile photo of \",[\"name\"]],\"+j94GR\":[\"About your hosting\"],\"/DXTzN\":[\"Please <0>start a new one</0>.\"],\"/NgYQS\":[\"Choose your Solid provider or WebID\"],\"/QPS+F\":[\"Open profile\"],\"/Rg4q2\":[\"Contacts of <0/>\"],\"/ZSU/m\":[\"Contact confirmed\"],\"/qBoT3\":[\"Create public type index:\"],\"01t3Qz\":[\"Sending message failed: \"],\"17JZAf\":[\"Create inbox:\"],\"1Lxtdt\":[\"code\"],\"1i+H23\":[\"Ignoring contact request\"],\"1rUvkC\":[\"Preparing Solid Pod\"],\"2FwEm8\":[\"Authenticating...\"],\"2LDW9H\":[\"contacts\"],\"2ZJVX8\":[\"logo of \",[\"name\"]],\"2n699w\":[\"Community does not have a group.\"],\"2r6kDW\":[\"Hospex data setup successful\"],\"3pqzUK\":[\"Contact added.\"],\"42aUkb\":[\"Enter your email address:\"],\"47l1wN\":[\"Create private type index:\"],\"49hNay\":[\"You can also <0>host your own Pod</0>.\"],\"4Wck3W\":[\"Request failed: \",[\"status\"],\" \",[\"statusText\"],\" \",[\"url\"]],\"5lWFkC\":[\"Sign in\"],\"5mmKJA\":[\"New conversation with \",[\"people\"]],\"65VN7K\":[\"Interests\"],\"6Kt8NA\":[\"Fetching profile\"],\"6YtxFj\":[\"Name\"],\"6dc9lc\":[\"Interest removed\"],\"6pe0RU\":[\"Session Expired!\"],\"7VpFf+\":[\"Your \",[\"communityLabel\"],\" data storage is set up.\"],\"7hktsm\":[\"Getting started\"],\"8BwQeU\":[\"Finish Setup\"],\"8kXsjC\":[\"Hi! I'd like to add you as a contact!\"],\"9K7sK1\":[\"Email notification was sent\"],\"9OpVZg\":[\"Edit profile\"],\"A3rd0M\":[\"Sending Linked Data Notification failed.\"],\"A94/5H\":[\"Solid Pod is prepared\"],\"ARpyUK\":[\"Creating channel failed\"],\"ApEC55\":[\"Hosting location (click <0/> or drag & zoom map)\"],\"AywWH9\":[\"Public type index is set up.\"],\"B8AaMI\":[\"This field is required\"],\"BJjAHC\":[\"my hosting\"],\"BLFgtR\":[\"Join Community\"],\"BQTO6T\":[\"For safety, share only approximate location\"],\"BajMQF\":[\"Send contact invitation\"],\"Bn8bAf\":[\"WIP\"],\"Bpp0Bz\":[\"Introduce yourself — it builds trust\"],\"C++TGh\":[\"Close hosting preview\"],\"CDTRqk\":[\"You are already a member of \",[\"communityLabel\"]],\"CMfhBr\":[\"Hospex not set up (should not happen)\"],\"CiQd9C\":[\"Add Accommodation\"],\"Ct2zYB\":[\"Joining community\"],\"DvnzIZ\":[\"You can only read this chat.\"],\"ER/wp5\":[\"Updating profile\"],\"EbjjWU\":[\"Confirming contact\"],\"EnJuK0\":[\"Conversations\"],\"F132R+\":[\"Please verify your email\"],\"FIq1Ba\":[\"Later\"],\"FQjzSF\":[\"Prepare Pod\"],\"GE4F4y\":[\"Start a new conversation\"],\"GbJuzd\":[[\"messageCount\",\"plural\",{\"one\":[\"(\",\"#\",\" new)\"],\"other\":[\"(\",\"#\",\" new)\"]}]],\"H2Pd5s\":[\"Preferences file is set up.\"],\"HBI6nY\":[\"Need help?\"],\"HFwLuM\":[\"This chat has outdated format.\"],\"HSNeM4\":[\"Community Solid Pod is not available at the moment.\"],\"HZ3V75\":[\"Sending Solid notification\"],\"HecDqj\":[\"edit profile\"],\"Hwv3lE\":[\"Sign in to \",[\"communityName\"]],\"I6Q2R9\":[\"Verification email sent\"],\"IAssyr\":[\"Please wait for everything to load.\"],\"IJbwnu\":[\"About me\"],\"IYqOgL\":[\"Please invite only people you know personally.\"],\"JlFcis\":[\"Send\"],\"JmZ/+d\":[\"Finish\"],\"KBHVBP\":[\"Join \",[\"communityName\"]],\"KhBSE/\":[\"Invitation\"],\"KuVpz+\":[\"Unread\"],\"LShGrM\":[\"focused logo of \",[\"name\"]],\"LpmI27\":[\"Adding contact.\"],\"M/R9wy\":[\"Creating message\"],\"MHrjPM\":[\"Title\"],\"MPWsQ5\":[\"Public type index is not set up\"],\"MYOoIS\":[\"A confirmation email has been sent to \",[\"sentToEmail\"],\".<0/>Please check your inbox (and spam folder) and follow the instructions.\"],\"N0mEtw\":[\"There doesn't seem to be a chat here.\"],\"NLJ18A\":[\"hospex container not found (too soon?)\"],\"NWQASA\":[\"Email Notification was sent.\"],\"NqP6cw\":[\"Messages with <0/>\"],\"O75yBm\":[\"loading interests...\"],\"OjgFFJ\":[\"Please move map to your hosting location\"],\"PS1p9R\":[\"Something went wrong.\\nError: \",[\"e\"]],\"PUl3lo\":[\"Accommodation updated\"],\"PfxFCG\":[\"No hospex document for community\"],\"PmxYBp\":[\"Sending email notification\"],\"Pygpzd\":[\"We didn't succeed with redirecting to the issuer at \",[\"oidcIssuer\"],\".\\nHave you provided correct webId or OIDCIssuer? Or is it down?\\n\\nReason: \",[\"message\"],\"\\n\\nError: \",[\"errorString\"]],\"Q1iVjv\":[\"e.g. solidcommunity.net\"],\"QlfR57\":[\"Accommodation created\"],\"QwanTV\":[\"Welcome to \",[\"communityLabel\"],\"!\"],\"Rad1+W\":[\"<0>Reach us in our project spaces</0> or write us email to <1>\\\"support at solidcouch dot org\\\"</1>.\"],\"S48xcO\":[\"pending\"],\"S8rY67\":[\"Removing interest\"],\"SH3fsl\":[\"Preparing and sending verification email\"],\"TFYtVq\":[\"Add person as a contact\"],\"TQ1cfc\":[\"Accommodation not found.\"],\"V6us8s\":[\"Linked Data Notification was sent.\"],\"VM9oSF\":[\"Your email notifications are set up.\"],\"VXHB7t\":[\"Find people to stay with\"],\"VhYqBG\":[\"Contact request from <0/>\"],\"WKVrQ3\":[\"Choose a Pod provider:\"],\"WO3k4H\":[\"process\"],\"Wixo6f\":[\"Do you really want to delete the accommodation?\"],\"XCH5Qf\":[\"Something is not resolved (unexpected error)\"],\"XQAIBA\":[\"Unexpected status code from indexing service\"],\"XuPkTz\":[\"Skip getting started\"],\"XwS+Aq\":[\"Link existing community profile:\"],\"Y8WO6Z\":[\"Adding \",[\"label\"],\" to interests\"],\"YVkYKz\":[\"Welcome travellers to your home — if you can\"],\"Z0DAcX\":[\"Accommodation removed from indexing service\"],\"Z3FXyt\":[\"Loading...\"],\"Z3PAZP\":[\"Your session has expired. Please sign in again.\"],\"ZDIydz\":[\"Get started\"],\"ZfmHL/\":[\"Choose your community data storage:\"],\"a8Qrhz\":[\"Something went wrong: \",[\"failure\"]],\"bC8dmZ\":[\"Create new profile for this community:\"],\"bGSJ3R\":[\"missing variables\"],\"bbvfbr\":[\"profile\"],\"bo2Xks\":[\"Loading previous messages\"],\"bpK9mh\":[\"Confirm and Continue\"],\"cnGeoo\":[\"Delete\"],\"d/3T+G\":[\"Inbox is set up.\"],\"dEgA5A\":[\"Cancel\"],\"dIY3v5\":[\"Community joined\"],\"dJsK6G\":[\"Contact exists\"],\"dWGab4\":[\"This is a start of your conversation\"],\"dg8/vM\":[\"Send a message…\"],\"e0IFR+\":[\"Switch to light theme\"],\"ePK91l\":[\"Edit\"],\"eWlz27\":[\"Write a message\"],\"ei1Uj7\":[\"Private type index is set up.\"],\"f7fqUX\":[\"Accommodation updated in indexing service\"],\"f8jrkd\":[\"more\"],\"fMnMgg\":[\"Private type index not set up (should not happen)\"],\"fo1jSY\":[\"Inbox is not set up\"],\"g3UF2V\":[\"Accept\"],\"gHFvtg\":[\"messages\"],\"gbbkE7\":[\"Ignore\"],\"hHo0hg\":[\"Accommodation offer from \",[\"offeredByIdentifier\"]],\"hXzOVo\":[\"Next\"],\"heRuNl\":[\"Email notifications are not available for this community.\"],\"hyxB5M\":[\"Join request sent\"],\"iKT85g\":[\"To join \",[\"communityName\"],\", you need a <0>Solid Pod</0> — a place to store your data.\"],\"iRVV2L\":[\"My Accommodation Offers\"],\"ioBnUo\":[\"No storage found\"],\"jD5ug7\":[\"Sending Email Notification failed.\"],\"jRqiIC\":[\"Contact request ignored\"],\"jevYpv\":[\"Create preferences file:\"],\"jjIi3z\":[\"Don't have a Solid Pod yet?\"],\"jlqRFZ\":[\"travel\"],\"jx1+4B\":[[\"label\"],\" added to interests\"],\"k4QqVn\":[\"Add \",[\"locale\"]],\"kH4vZi\":[\"Message was created\"],\"kH9S1u\":[\"Join community \",[\"communityLabel\"]],\"kf83Ld\":[\"Something went wrong.\"],\"lYCAyX\":[\"Setup email notifications\"],\"maeAdU\":[\"fetching contacts\"],\"meH4FD\":[\"Add to my contacts\"],\"mkeWa1\":[\"\\\"\",[\"channel\"],\"\\\" is not a valid chat channel\"],\"nwtY4N\":[\"Something went wrong\"],\"ocwjau\":[\"email address\"],\"pCyggK\":[\"Chat channel was created\"],\"pY95XS\":[\"<0/> wrote you:\"],\"piu7ZA\":[\"See contact invitation\"],\"qFkBv/\":[\"Solid notification was sent\"],\"qPXQY8\":[\"authenticating\"],\"qa9cy4\":[\"Photo\"],\"r60ENi\":[\"Creating accommodation\"],\"sCXkJv\":[\"Set Up Notifications\"],\"sRGBFr\":[\"inbox not found (too soon?)\"],\"sa0+P5\":[\"Select my location\"],\"srPuxS\":[\"Profile updated\"],\"t9VobG\":[\"Accommodation deleted\"],\"tEB6c1\":[\"host\"],\"tErrdj\":[\"Conversations with \",[\"people\"]],\"tfDRzk\":[\"Save\"],\"uLvjiZ\":[\"Add a language\"],\"udP0jg\":[\"No authenticated user available\"],\"uliB7J\":[\"Resend Confirmation Email in \",[\"countdown\"],\"s\"],\"uwiVz0\":[\"Contact request sent\"],\"v5eALp\":[\"My location\"],\"vJIeDW\":[\"Tell others about your place and hosting\"],\"vTkuq4\":[\"Notifying indexing service\"],\"vVEMd1\":[\"Hospex document not found\"],\"vp9yIB\":[\"Send Confirmation Email\"],\"vz3tpL\":[\"Accommodation added to indexing service\"],\"wRoTH5\":[\"Messages with \",[\"personLabel\"]],\"xAkrnZ\":[\"Inbox of the other person not found (probably too soon)\"],\"xGVfLh\":[\"Continue\"],\"xdx7Sc\":[\"Updating accommodation\"],\"y3aU20\":[\"Save changes\"],\"yNR1Vx\":[\"Switch to dark theme\"],\"yQE2r9\":[\"Loading\"],\"yXo/Y0\":[\"Powered by <0>SolidCouch</0> v\",[\"version\"],\" (<1>\",[\"commitHashShort\"],\"</1>)\"],\"ydgdD7\":[\"sign out\"],\"yz7wBu\":[\"Close\"],\"z8F4/3\":[\"Your email\"],\"zZiCw4\":[\"Setting up hospex data\"]}")as Messages;
