+++ src/pages/Profile/ProcessContactRequest.module.scss
@@ -1,0 +1,26 @@
+@use '@/styles/media';
+
+.contactRequestModal {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4x);
+
+  // TODO duplicate AddContactForm.module.scss
+  .actions {
+    display: flex;
+    align-items: stretch;
+    gap: var(--space-4x);
+
+    button {
+      flex: 1;
+    }
+
+    @include media.for-tablet-portrait-up {
+      justify-content: flex-end;
+
+      button {
+        flex: initial;
+      }
+    }
+  }
+}
